requirements.txt
django

# For work with postgresql
psycopg2

# For debug
ipdb

# For work with images
pillow
